requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymongo==3.12.3
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
Pillow==11.0.0
python-decouple==3.8
daphne==4.0.0
asgiref==3.7.2
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
